Add multi-version operation to freezer (#277)
* Fix the problem that rpm cannot join the plugins .so

* Add multi-version operation to freezer

* Fix the bug caused by missing version value

* Fix bugs caused by branch merge

+++ storage/multipart.go
@@ -582,7 +582,7 @@ func (yig *YigStorage) CompleteMultipartUpload(reqCtx RequestContext, credential
 		CreateTime:       uint64(now.UnixNano()),
 	}
 	object.VersionId = object.GenVersionId(bucket.Versioning)
-	if object.StorageClass == ObjectStorageClassGlacier {
+	if object.StorageClass == ObjectStorageClassGlacier && bucket.Versioning != datatype.BucketVersioningEnabled {
 		freezer, err := yig.MetaStorage.GetFreezer(object.BucketName, object.Name, object.VersionId)
 		if err == nil {
 			err = yig.MetaStorage.DeleteFreezer(freezer)
