+++ storage/object.go
@@ -622,7 +622,7 @@ func (yig *YigStorage) PutObject(reqCtx RequestContext, credential common.Creden
 		CreateTime:           uint64(now.UnixNano()),
 	}
 	object.VersionId = object.GenVersionId(bucket.Versioning)
-	if object.StorageClass == ObjectStorageClassGlacier {
+	if object.StorageClass == ObjectStorageClassGlacier && bucket.Versioning != datatype.BucketVersioningEnabled {
 		freezer, err := yig.MetaStorage.GetFreezer(object.BucketName, object.Name, object.VersionId)
 		if err == nil {
 			err = yig.MetaStorage.DeleteFreezer(freezer)
@@ -922,7 +922,7 @@ func (yig *YigStorage) removeOldObject(object *meta.Object) (err error) {
 	}
 
 	if object.StorageClass == ObjectStorageClassGlacier {
-		freezer, err := yig.GetFreezer(object.BucketName, object.Name, "")
+		freezer, err := yig.GetFreezer(object.BucketName, object.Name, object.VersionId)
 		if err == nil {
 			if freezer.Name == object.Name {
 				err = yig.MetaStorage.DeleteFreezer(freezer)
